+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - On X11, fix `ResumeTimeReached` being fired too early.
 - On Web, replaced zero timeout for `ControlFlow::Poll` with `requestAnimationFrame`
 - On Web, fix a possible panic during event handling
+- On Wayland, add missing virtual key codes mappings
 
 # 0.22.0 (2020-03-09)